Update README.md with overview content
Incorrectly put this text content into the wrong chapter - it belongs here instead. I want to test if uploading this text to the README.md causes it to appear in the section title page in the GitBook. If so, we don't need separate overview pages in each folder, and can instead upload our overview text directly to the README.md files in each folder.