lateral reading
The best way to establish the credibility of information is through a process called **lateral reading**. The process of lateral reading involves reading across sources, checking information as it’s presented to you. This contrasts from simply reading a webpage's 'About Us' page and taking it at face value, to also searching the rest of the web for more information about the author or organisation who created the information, the date it was posted, what others have to say about the topic and more.  As an evaluation method, we recommend the CRAAP method, which checks for Currency, Relevance, Authority, Accuracy and Purpose.